refactor: move notion use cases to module
